Доделать изменение строки В changeColumn всё готово

+++ TheGUIAndSQLApp/MainWindow.xaml.cs
@@ -30,22 +30,26 @@ namespace TheGUIAndSQLApp
         bool tableSelected = false;
         public string currentDataBase = "";
         string currentTable = "";
+
         public MainWindow()
         {
             InitializeComponent();
             table.Items.Clear();
             dataBases.Items.Clear();
             openedTableGrid.Items.Clear();
+            describeTable.Items.Clear();
         }
 
         private void CloseCustomDialog(object sender, RoutedEventArgs e)
         {
             CloseCustomDialogRealize(1);
         }
+
         private void CloseCustomDialog1(object sender, RoutedEventArgs e)
         {
             CloseCustomDialogRealize(2);
         }
+
         private async void CloseCustomDialogRealize(int id)
         {
             string resource = "";
@@ -62,38 +66,10 @@ namespace TheGUIAndSQLApp
 
         private void UsingSql(object sender, RoutedEventArgs e)
         {
-            string sConn = @"Database = EmployeesBase; Data Source = localhost; User Id = root; Password =";
-            MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
-            result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset(this.SQLQuery.Text, sConn);
-
-            //MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(this.SQLQuery.Text, "Data Source=localhost;User Id=root;Password=q13466431");
-              
-             if (result.HasError == false)
-             {
-                table.ItemsSource = result.ResultData.DefaultView;
-                for(int i = 0; i < table.Columns.Count; i++)
-                {
-                    if (table.Columns[i].Header.ToString() == "photopath")
-                    {
-                        table.Columns[i].Visibility = Visibility.Hidden;
-                    }
-                }
-            }
-             else
-             {
-                 MessageBox.Show(result.ErrorText);
-             }
-        }
-
-        private void table_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            string sConn = @"Database = EmployeesBase; Data Source = localhost; User Id = root; Password =";
-            MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
-            result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SELECT photopath FROM Employees WHERE ID = " + (table.SelectedIndex+1).ToString() + ";", sConn);
 
         }
 
-        private void Button_Click(object sender, RoutedEventArgs e)
+        private void Auth_Click(object sender, RoutedEventArgs e)
         {
             AuthGrid.Visibility = Visibility.Collapsed;
             MainGrid.Visibility = Visibility.Visible;
@@ -104,9 +80,6 @@ namespace TheGUIAndSQLApp
             string sConn = @"Database = EmployeesBase; Data Source = localhost; User Id = root; Password =";
             MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
             result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SELECT * FROM Employees", sConn);
-
-            //MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(this.SQLQuery.Text, "Data Source=localhost;User Id=root;Password=q13466431");
-
             if (result.HasError == false)
             {
                 table.ItemsSource = result.ResultData.DefaultView;
@@ -121,9 +94,6 @@ namespace TheGUIAndSQLApp
             string sConn = @"Database = EmployeesBase; Data Source = localhost; User Id = root; Password =";
             MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
             result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SHOW DATABASES", sConn);
-
-            //MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(this.SQLQuery.Text, "Data Source=localhost;User Id=root;Password=q13466431");
-
             if (result.HasError == false)
             {
                 dataBases.ItemsSource = result.ResultData.DefaultView;
@@ -137,29 +107,43 @@ namespace TheGUIAndSQLApp
 
         private void DataBase_Open(object sender, RoutedEventArgs e)
         {
-            string str = ((DataRowView)dataBases.SelectedItems[0]).Row["DATABASE"].ToString();
-            string sConn = @"Database = " + str + "; Data Source = localhost; User Id = root; Password =";
-            MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
-            result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SHOW TABLES", sConn);
-            if (result.HasError == false)
+            if (dataBases.SelectedItem == null)
             {
-                dataBaseSelected = true;
-                currentDataBase = str;
-                tabControl.SelectedIndex = 1;
-                tabControl.SelectedItem = tablesTab;
-                tablesTab.IsSelected = true;
-                table.ItemsSource = result.ResultData.DefaultView;
+                somethingAintSelected("База данных не выбрана", "Выберите базу данных, которую хотите открыть");
             }
             else
             {
-                MessageBox.Show(result.ErrorText);
+                string str = ((DataRowView)dataBases.SelectedItems[0]).Row["DATABASE"].ToString();
+                string sConn = @"Database = " + str + "; Data Source = localhost; User Id = root; Password =";
+                MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
+                result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SHOW TABLES", sConn);
+                if (result.HasError == false)
+                {
+                    dataBaseSelected = true;
+                    currentDataBase = str;
+                    tabControl.SelectedIndex = 1;
+                    tabControl.SelectedItem = tablesTab;
+                    tablesTab.IsSelected = true;
+                    table.ItemsSource = result.ResultData.DefaultView;
+                }
+                else
+                {
+                    MessageBox.Show(result.ErrorText);
+                }
             }
         }
 
         private void DataBase_Delete(object sender, RoutedEventArgs e)
         {
-            string str = ((DataRowView)dataBases.SelectedItems[0]).Row["DATABASE"].ToString();
-            showAcceptDropDBDialog("DROP DATABASE `" + str + "`");
+            if (dataBases.SelectedItem == null)
+            {
+                somethingAintSelected("База данных не выбрана", "Выберите базу данных, которую хотите удалить");
+            }
+            else
+            {
+                string str = ((DataRowView)dataBases.SelectedItems[0]).Row["DATABASE"].ToString();
+                showAcceptDropDBDialog("DROP DATABASE `" + str + "`");
+            }
         }
 
         private void TabItem_Loaded(object sender, RoutedEventArgs e)
@@ -173,9 +157,24 @@ namespace TheGUIAndSQLApp
             {
                 showChooseDialog(1);
             }
-            if (tableSelected == false && tabControl.SelectedIndex == 2)
+            if (tableSelected == false && (tabControl.SelectedIndex == 2 ||
+                tabControl.SelectedIndex == 3))
             {
                 showChooseDialog(2);
+            }
+            switch (tabControl.SelectedIndex)
+            {
+                case 2:
+                    statusBar.Visibility = Visibility.Visible;
+                    statusBarType.Content = "Содержание таблицы";
+                    statusBarSelected.Content = currentTable;
+                    break;
+                case 3:
+                    statusBar.Visibility = Visibility.Visible;
+                    statusBarType.Content = "Структура таблицы";
+                    statusBarSelected.Content = currentTable;
+                    break;
+                default: statusBar.Visibility = Visibility.Collapsed; break;
             }
         }
         private async void showChooseDialog(int type)
@@ -216,7 +215,7 @@ namespace TheGUIAndSQLApp
             }
         }
 
-            private async void showAcceptDropDBDialog(string dropStr)
+            private async void showAcceptDropDBDialog(string queryStr)
             {
             var mySettings = new MetroDialogSettings()
             {
@@ -225,22 +224,32 @@ namespace TheGUIAndSQLApp
                 ColorScheme = MetroDialogOptions.ColorScheme
             };
 
-            MessageDialogResult result = await this.ShowMessageAsync("Вы действительно хотите выполнить запрос?", dropStr,
+            MessageDialogResult result = await this.ShowMessageAsync("Вы действительно хотите выполнить запрос?", queryStr,
                 MessageDialogStyle.AffirmativeAndNegative, mySettings);
 
             if(result == MessageDialogResult.Affirmative)
             {
-                if(dropStr.Contains("DROP TABLE"))
+                if(queryStr.Contains("DROP TABLE `") || queryStr.Contains("TRUNCATE `"))
                 {
-                    MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(dropStr, "Database = " + currentDataBase + "; Data Source=localhost;User Id=root;Password=");
+                    MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(queryStr, "Database = " + currentDataBase + "; Data Source=localhost;User Id=root;Password=");
+                    if(tabControl.SelectedIndex == 2 || queryStr.Contains(currentTable))
+                    {
+                        tabControl.SelectedIndex = 1;
+                        tableSelected = false;
+                    }
                     updateCurrentTable();
                 }
                 else
                 {
-                    MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(dropStr, "Data Source=localhost;User Id=root;Password=");
+                    MySqlLib.MySqlData.MySqlExecute.SqlNoneQuery(queryStr, "Data Source=localhost;User Id=root;Password=");
+                    if(queryStr.Contains(currentDataBase))
+                    {
+                        tabControl.SelectedIndex = 0;
+                        tableSelected = false;
+                        dataBaseSelected = false;
+                    }
                     SelectDataBases();
                 }
-               
             }
         }
 
@@ -260,22 +269,30 @@ namespace TheGUIAndSQLApp
 
         private void Table_Open(object sender, RoutedEventArgs e)
         {
-            string str = ((DataRowView)table.SelectedItems[0]).Row[0].ToString();
-            string sConn = @"Database = " + currentDataBase + "; Data Source = localhost; User Id = root; Password =";
-            MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
-            result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SELECT * FROM "+str, sConn);
-            if (result.HasError == false)
+            if (table.SelectedItem == null)
             {
-                tableSelected = true;
-                currentTable = str;
-                tabControl.SelectedIndex = 2;
-                tabControl.SelectedItem = openedTable;
-                openedTable.IsSelected = true;
-                openedTableGrid.ItemsSource = result.ResultData.DefaultView;
+                somethingAintSelected("Таблица не выбрана", "Выберите таблицу, которую хотите открыть");
             }
             else
             {
-                MessageBox.Show(result.ErrorText);
+                string str = ((DataRowView)table.SelectedItems[0]).Row[0].ToString();
+                string sConn = @"Database = " + currentDataBase + "; Data Source = localhost; User Id = root; Password =";
+                MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
+                result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SELECT * FROM " + str, sConn);
+                if (result.HasError == false)
+                {
+                    tableSelected = true;
+                    currentTable = str;
+                    tabControl.SelectedIndex = 2;
+                    tabControl.SelectedItem = openedTable;
+                    openedTable.IsSelected = true;
+                    openedTableGrid.ItemsSource = result.ResultData.DefaultView;
+                    describeTableFoo();
+                }
+                else
+                {
+                    MessageBox.Show(result.ErrorText);
+                }
             }
         }
 
@@ -306,11 +323,98 @@ namespace TheGUIAndSQLApp
                 MessageBox.Show(result.ErrorText);
             }
         }
+
         private void Table_Delete(object sender, RoutedEventArgs e)
         {
-            string rowStr = "TABLESIN_" + currentDataBase.ToUpper();
-            string str = ((DataRowView)table.SelectedItems[0]).Row[0].ToString();
-            showAcceptDropDBDialog("DROP TABLE `" + str + "`");
+            if (table.SelectedItem == null)
+            {
+                somethingAintSelected("Таблица не выбрана", "Выберите таблицу, которую хотите удалить");
+            }
+            else
+            {
+                string str = ((DataRowView)table.SelectedItems[0]).Row[0].ToString();
+                showAcceptDropDBDialog("DROP TABLE `" + str + "`");
+            }
+        }
+
+        private void showStructure(object sender, RoutedEventArgs e)
+        {
+            describeTableFoo();
+            tabControl.SelectedIndex = 3;
+        }
+        private void describeTableFoo()
+        {
+            if (tableSelected == false)
+            {
+                showChooseDialog(2);
+            }
+            else
+            {
+                string sConn = @"Database = " + currentDataBase + "; Data Source = localhost; User Id = root; Password =";
+                MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
+                result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("DESCRIBE `" + currentTable + "`", sConn);
+                if (result.HasError == false)
+                {
+                    describeTable.ItemsSource = result.ResultData.DefaultView;
+                }
+                else
+                {
+                    MessageBox.Show(result.ErrorText);
+                }
+            }
+        }
+
+        private void truncateTable(object sender, RoutedEventArgs e)
+        {
+            showAcceptDropDBDialog("TRUNCATE `" + currentTable + "`");
+        }
+
+        private void updateTable(object sender, RoutedEventArgs e)
+        {
+            string sConn = @"Database = " + currentDataBase + "; Data Source = localhost; User Id = root; Password =";
+            MySqlLib.MySqlData.MySqlExecuteData.MyResultData result = new MySqlLib.MySqlData.MySqlExecuteData.MyResultData();
+            result = MySqlLib.MySqlData.MySqlExecuteData.SqlReturnDataset("SELECT * FROM " + currentTable, sConn);
+            if (result.HasError == false)
+            {
+                openedTableGrid.ItemsSource = result.ResultData.DefaultView;
+            }
+            else
+            {
+                MessageBox.Show(result.ErrorText);
+            }
+        }
+
+        private void deleteTable(object sender, RoutedEventArgs e)
+        {
+            showAcceptDropDBDialog("DROP TABLE `" + currentTable + "`");
+        }
+
+
+        private void changeTable(object sender, RoutedEventArgs e)
+        {
+            if(openedTableGrid.SelectedItem == null)
+            {
+                somethingAintSelected("Строка не выбрана", "Сначала выберите строку, которую хотите изменить");
+            }
+            else
+            {
+                string[] rows = new string[openedTableGrid.Columns.Count];
+                for (int i = 0; i < openedTableGrid.Columns.Count; i++)
+                {
+                    rows[i] = ((DataRowView)openedTableGrid.SelectedItem).Row[i].ToString();
+                }
+            }
+        }
+
+        private void somethingAintSelected(string caption, string text)
+        {
+            var mySettings = new MetroDialogSettings()
+            {
+                AffirmativeButtonText = "OK",
+                ColorScheme = MetroDialogOptions.ColorScheme
+            };
+
+            this.ShowMessageAsync(caption, text, MessageDialogStyle.Affirmative, mySettings);
         }
     }
 }
